bindings/swift: delete legacy hand-written types (step 4)
Remove OuisyncClient, OuisyncError, OuisyncFile, OuisyncEntry,
OuisyncRepository, OuisyncMessage, OuisyncLib, and OuisyncSession —
all superseded by generated Api.swift types and the new Client actor.
Extract NotificationStream into Sources/ since it has no FFI deps.
OuisyncFFI.swift is kept as a stub pending step 5.
